Switched to hard deletion for files and folders
The code changes involve switching from soft deletion to hard deletion in the MediaCategoryController. This was done by replacing calls to DeleteFilesSoft and DeleteFoldersSoft with DeleteFilesHard and DeleteFoldersHard respectively. Additionally, a new method called DeleteFoldersHard was implemented in FileAdminService which deletes folders and their associated files permanently. The IFileAdminService interface was also updated to include this new method. These changes provide a more thorough cleanup of resources when needed.